pyproject.toml
[project]
name = "pucoti"
version = "0.12.5"
description = "A Purposeful Countdown Timer"
readme = "README.md"
license = { text = "GPL-3.0" }
keywords = ["timer", "countdown", "pygame", "gui"]
requires-python = ">=3.12"
dependencies = [
    "desktop-notifier>=6.1.1",
    "luckypot>=0.1.1",
    "platformdirs>=4.3.6",
    "pydantic>=2.9.2",
    "pygame-ce>=2.5.2",
    "pyobjc-framework-cocoa==10.3.1 ; sys_platform == 'darwin'",
    "pyyaml>=6.0.2",
    "pyzmq>=26.2.0",
    "requests>=2.32.3",
    "ruamel-yaml>=0.18.6",
    "typer>=0.13.0",
    "umami-analytics>=0.2.20",
]

[dependency-groups]
server = [
    "fastapi>=0.115.4",
    "uvicorn>=0.32.0",
]
dev = [
    "ipykernel>=6.29.5",
    "pre-commit>=4.0.1",
    "pytest>=8.3.3",
]


[project.urls]
Homepage = "https://github.com/ddorn/pucoti"
Repository = "https://github.com/ddorn/pucoti"

[project.scripts]
pucoti = "pucoti.app:app"
pucoti-msg = "pucoti.controller:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
pythonpath = ["src"]
